Update testing and plugin codes and bump package version
Made updates in the testing codes, replacing `no_subprocess_run` fixture with `mock_run`. This change allows a finer control on the created mock, enabling a better testing environment. Also, modified some methods in the plugin.py file, changing the way the method `copy_what_new_does` worked. It is now renamed to `read_location` and has improved performance. To reflect the modifications in the code, a version bump to 0.1.0 was also made in the __about__.py file.