Add Python 3.10 SDK details for MacOS environment
Added the SDK description for Python 3.10 in the mac_os_jdk.tools.xml file, as well as the system configuration for MacOS in the macos_sys_config.json file. This ensures that the environment is correctly set up for Python 3.10 on MacOS, necessary for supporting the recent update to this version of Python.